% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_gap.R
\name{round_gap}
\alias{round_gap}
\alias{floor_gap}
\alias{ceiling_gap}
\title{Round to a Gap of Given Width}
\usage{
floor_gap(x, width)

ceiling_gap(x, width)
}
\arguments{
\item{x}{numeric or integer vector}

\item{width}{defines how wide a 'gap' is}
}
\description{
Round to a Gap of Given Width
}
\examples{
floor_gap(c(5, 7, 8, 11, 15, 17, 25), width = 3)
ceiling_gap(c(5, 7, 8, 11, 15, 17, 25), width = 3)
}
