% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_gap}
\alias{round_gap}
\alias{floor_gap}
\alias{ceiling_gap}
\title{Round to a Gap of Given Width}
\usage{
floor_gap(x, width)

ceiling_gap(x, width)
}
\arguments{
\item{x}{numeric or integer vector}

\item{width}{defines how wide a 'gap' is}
}
\description{
Round to a Gap of Given Width
}
